PR changes and config
