+++ src/config.py
@@ -16,6 +16,7 @@ class ProdConfig(Config):
 	FLASK_ENV = 'production'
 	DEBUG = False
 	TESTING = False
+	SECRET_KEY = os.environ.get('SECRET_KEY')
 
 	SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URI')
 
@@ -23,6 +24,7 @@ class DevConfig(Config):
 	FLASK_ENV = 'development'
 	DEBUG = True
 	TESTING = True
+	SECRET_KEY = os.environ.get('SECRET_KEY')
 
 	SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URI')
 
@@ -30,6 +32,7 @@ class LocalConfig(Config):
 	FLASK_ENV = 'local'
 	DEBUG = True
 	TESTING = True
+	SECRET_KEY='skbGoMfmaQzJTM'
 
 	SQLALCHEMY_DATABASE_URI = 'postgresql+psycopg2://postgres:password@localhost/postgres'
 	SQLALCHEMY_ENGINE_OPTIONS = {
@@ -42,5 +45,6 @@ class TestConfig(Config):
 	FLASK_ENV = 'test'
 	DEBUG = True
 	TESTING = True
+	SECRET_KEY='skbGoMfmaQzJTM'
 
 	SQLALCHEMY_DATABASE_URI = 'sqlite:///:memory:'
